Create modules for app and fix linter errors
